Fix redirects with authRequest/Success/Error action.
Before this, refreshing the auth token would redirect to /dashboard,
even if the initial URL requested was /something-else.

Fixed by introducing the three sync actions for authentication, and an
HoC, requiresLogin.  If the HoC sees that we are still waiting for auth,
it displays a "Logging in" message, and only redirects when it knows we
have failed to log in.